lots of changes

+++ database/migrations/2020_05_20_000000_create_media_table.php
@@ -4,17 +4,19 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class CreateMediaTable extends Migration {
+class CreateMediaTable extends Migration
+{
     /**
      * Run the migrations.
      *
      * @return void
      */
-    public function up() {
+    public function up()
+    {
         Schema::create(
             'media',
-            function(Blueprint $table) {
-                $table->increments('id');
+            function (Blueprint $table) {
+                $table->bigIncrements('id');
                 $table->string('title');
                 $table->string('filename');
                 $table->string('crc', 32);
@@ -29,7 +31,8 @@ class CreateMediaTable extends Migration {
      *
      * @return void
      */
-    public function down() {
+    public function down()
+    {
         Schema::dropIfExists('media');
     }
 }
