+++ database/migrations/2020_08_09_000000_create_upload_table.php
@@ -4,17 +4,19 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class CreateUploadTable extends Migration {
+class CreateUploadTable extends Migration
+{
     /**
      * Run the migrations.
      *
      * @return void
      */
-    public function up() {
+    public function up()
+    {
         Schema::create(
             'uploads',
-            function(Blueprint $table) {
-                $table->increments('id');
+            function (Blueprint $table) {
+                $table->bigIncrements('id');
                 $table->string('filename');
                 $table->foreignId('user_id')->constrained('users', 'id')->cascadeOnDelete();
                 $table->smallInteger('status')->unsigned();
@@ -27,7 +29,8 @@ class CreateUploadTable extends Migration {
      *
      * @return void
      */
-    public function down() {
+    public function down()
+    {
         Schema::dropIfExists('uploads');
     }
 }
